Source/mya/Source/Library/mya-catk/src/main/java/com/philips/platform/mya/catk/dto/CreateConsentDto.java
/*
 * Copyright (c) 2017 Koninklijke Philips N.V.
 * All rights are reserved. Reproduction or dissemination
 * in whole or in part is prohibited without the prior written
 * consent of the copyright holder.
 */

package com.philips.platform.mya.catk.dto;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

public class CreateConsentDto {

    @SerializedName("resourceType")
    @Expose
    private String resourceType;
    @SerializedName("language")
    @Expose
    private String language;
    @SerializedName("status")
    @Expose
    private String status;
    @SerializedName("subject")
    @Expose
    private String subject;
    @SerializedName("policyRule")
    @Expose
    private String policyRule;

    public CreateConsentDto() {
    }

    public CreateConsentDto(String language, String policyRule, String resourceType, String status, String subject) {
        this.resourceType = resourceType;
        this.language = language;
        this.status = status;
        this.subject = subject;
        this.policyRule = policyRule;
    }

    public String getResourceType() {
        return resourceType;
    }

    public void setResourceType(String resourceType) {
        this.resourceType = resourceType;
    }

    public String getLanguage() {
        return language;
    }

    public void setLanguage(String language) {
        this.language = language;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public String getSubject() {
        return subject;
    }

    public void setSubject(String subject) {
        this.subject = subject;
    }

    public String getPolicyRule() {
        return policyRule;
    }

    public void setPolicyRule(String policyRule) {
        this.policyRule = policyRule;
    }

}